Remove old floating "add element to chat" overlay
