+++ src/vs/workbench/workbench.web.main.ts
@@ -70,7 +70,6 @@ import './services/userDataProfile/browser/userDataProfileStorageService.js';
 import './services/configurationResolver/browser/configurationResolverService.js';
 import '../platform/extensionResourceLoader/browser/extensionResourceLoaderService.js';
 import './services/auxiliaryWindow/browser/auxiliaryWindowService.js';
-import './services/browserElements/browser/webBrowserElementsService.js';
 import './services/power/browser/powerService.js';
 import '../platform/sandbox/browser/sandboxHelperService.js';
